fixed converted int 16 to int64 in weather property

+++ Weather/Model/Weather + Convenience.swift
@@ -5,12 +5,12 @@ extension Weather {
     
     convenience init?(location: Location, dictionary:[String:Any], context: NSManagedObjectContext = CoreDataStack.context) {
         self.init(context: context)
-        guard let timeStamp = dictionary[WeatherKeys.timeStamp.rawValue] as? Int16,
+        guard let timeStamp = dictionary[WeatherKeys.timeStamp.rawValue] as? Int,
             let maxTemp = dictionary[WeatherKeys.maxTemp.rawValue] as? Double,
             let minTemp = dictionary[WeatherKeys.minTemp.rawValue] as? Double else { return nil }
         
         self.location = location
-        self.timeStamp = timeStamp
+        self.timeStamp = Int64(timeStamp)
         self.maxTemp = maxTemp
         self.minTemp = minTemp
     }
